Refactor typo, alignment of table, and orderby
Fixed a typo (who's --> whose)

Fixed table alignment (add a space where one was missing and
updated the padding number accordingly)

Update the itertools.orderby() because I'd skipped over the
part of the docs where they state:
Generally, the iterable needs to already be sorted
on the same key function. The operation of
groupby() is similar to the uniq filter in Unix.
It generates a break or new group every time the
value of the key function changes (which is why it
is usually necessary to have sorted the data using
the same key function).
Previously I was passing BILL_ITEMS unsorted into the groupby(), which could cause incorrect grouping if the input order changed at any point.

Docs: https://docs.python.org/3/library/itertools.html#itertools.groupby